Squashed commit of the following:
commit ad47beb858bbc9caa1d9f57e2094d4e93514cd1e
Author: GregJohnsonJr <gregjohn456@gmail.com>
Date:   Wed Sep 11 15:20:23 2024 -0400

    Cleaning up build notes.

commit e7d8625309f771d6afbe6a185080ee98280e3dc9
Author: GregJohnsonJr <gregjohn456@gmail.com>
Date:   Wed Sep 11 14:58:20 2024 -0400

    Fix for negative index value

commit a44542181ba799cdd243a116b24d881e5dfed9b6
Author: Gregory Johnson <31873199+GregJohnsonJr@users.noreply.github.com>
Date:   Wed Sep 11 14:03:37 2024 -0400

    Fix compilation warnings (#12)

    * No warnings on mac!

    * Still fixing clang-tidy errors. Lots of comparisons between longs and ints that will cause problems.

    * Half way done, modified a lot of deduced type conversions. May need to go back in the future to remove them.

    * Should fix all the warnings on linux and windows.

    * Pushing through more deduced type fixes.

    * Should be the final fix for linux.

commit 2be797e4aa0b7746e0e845379a4d371e135ea2d9
Author: YoungKrug <gregjohn456@gmail.com>
Date:   Tue Sep 10 11:52:34 2024 -0400

    Removing and fixing check issues.

commit 3bd7dea6ceb30c43dd5cf2eee399d0445df8d6e2
Author: Gregory Johnson <31873199+GregJohnsonJr@users.noreply.github.com>
Date:   Tue Sep 10 11:27:47 2024 -0400

    Fix results (#11)

    * Adding example data for mothur results.

    * Added support for similarity matrix functionality

    * Updates to phylip writing, and fixes to incorrect clustering!

    * Stable, Classic cluster and opticluster both are identical to that of mothur! + Names can be added

    * Results are the same as they are in mothur!

    * All the unit test work, and clustur prints out expected results!

commit 7564d48b2258a9b09b0984c57d8524c10d73d457
Author: Gregory Johnson <31873199+GregJohnsonJr@users.noreply.github.com>
Date:   Tue Sep 3 11:21:42 2024 -0400

    Forgot a unit test. (#10)

+++ src/ListVectorTestFixture.cpp
@@ -54,7 +54,7 @@ bool ListVectorTestFixture::TestListVectorReturnsCorrectGetLabelsValue(
     for (const auto& sequences: mockListOfSequences) {
         listVector->push_back(sequences);
     }
-    const bool result = listVector->getLabels().size() == expectedResult;
+    const bool result = static_cast<int>(listVector->getLabels().size()) == expectedResult;
     TearDown();
     return result;
 }
@@ -74,7 +74,7 @@ bool ListVectorTestFixture::TestListVectorSetsLabelsCorrectly(const std::vector<
     const int expectedResult) {
     Setup();
     listVector->setLabels(mockListOfLabels);
-    const bool result = listVector->getLabels().size() == expectedResult;
+    const bool result = static_cast<int>(listVector->getLabels().size()) == expectedResult;
     TearDown();
     return result;
 }
@@ -119,7 +119,7 @@ bool ListVectorTestFixture::TestListVectorReturnsCorrectGetOtuNamesSize(const st
     for (const auto& sequences: mockListOfSequences) {
         listVector->push_back(sequences);
     }
-    const bool result = listVector->getOTUName(binToTest).size() == expectedResult;
+    const bool result = static_cast<int>(listVector->getOTUName(binToTest).size()) == expectedResult;
     TearDown();
     return result;
 }
